pyproject.toml
[project]
name = "vibecheck"
version = "0.2.0"
description = "基于 LLM 语义评语与 TF-IDF 的混合音乐推荐系统"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.124.4",
    "uvicorn>=0.38.0",
    "pydantic-settings>=2.0.0",
    "sqlalchemy>=2.0.45",
    "psycopg2-binary>=2.9.11",
    "pgvector>=0.4.2",
    "openai>=2.12.0",
    "requests>=2.32.5",
    "python-dotenv>=1.2.1",
    "jieba>=0.42.1",
    "scikit-learn>=1.8.0",
]

[project.scripts]
dev = "uvicorn app.main:app --reload --host 0.0.0.0 --port 8000"
